Organizing the tests/src/JIT/HardwareIntrinsics/X86 folder by ISA to make it more manageable.
