+++ tests/src/JIT/HardwareIntrinsics/X86/Lzcnt/Lzcnt.cs
