+++ tests/src/JIT/HardwareIntrinsics/X86/Lzcnt/Lzcnt_r.csproj
